// Plataforma-Secundaria/src/app/interfaces/alumno.ts
export interface Alumno{
    id?: number;
    nombre: string;
    apellido_Paterno: string;
    apellido_Materno: string;
    contraseña: string;
    grado: number;
    taller: string;
    grupo: string;

    // constructor(nombre: string, apellidoPaterno: string, apellidoMaterno: string, contraseña: string, grado: number, taller: string, grupo: string){
    //     this.nombre = nombre;
    //     this.apellidoPaterno = apellidoPaterno;
    //     this.apellidoMaterno	 = apellidoMaterno	;
    //     this.contraseña = contraseña;
    //     this.grado = grado;
    //     this.taller = taller;
    //     this.grupo = grupo;        
    // }
    
}